set _startTime/_finishTime to float64

+++ output/bulk_http.go
@@ -253,7 +253,7 @@ func (p *HTTPBulkProcessor) bulk(bulkRequest BulkRequest, execution_id int) {
 }
 
 func (p *HTTPBulkProcessor) innerBulk(bulkRequest BulkRequest, execution_id int) {
-	_startTime := float32(time.Now().UnixNano()/1000000) / 1000
+	_startTime := float64(time.Now().UnixNano()/1000000) / 1000
 	eventCount := bulkRequest.eventCount()
 	glog.Infof("bulk %d docs with execution_id %d", eventCount, execution_id)
 	for {
@@ -269,7 +269,7 @@ func (p *HTTPBulkProcessor) innerBulk(bulkRequest BulkRequest, execution_id int)
 		url := host
 		success, shouldRetry, noRetry, newBulkRequest := p.tryOneBulk(url, bulkRequest)
 		if success {
-			_finishTime := float32(time.Now().UnixNano()/1000000) / 1000
+			_finishTime := float64(time.Now().UnixNano()/1000000) / 1000
 			timeTaken := _finishTime - _startTime
 			glog.Infof("bulk done with execution_id %d %f %d %f", execution_id, timeTaken, eventCount, timeTaken/float32(eventCount))
 			p.hostSelector.addWeight(host)
